+++ src/main/java/chess/controller/SpringChessController.java
@@ -87,19 +87,4 @@ public class SpringChessController {
     public void deleteRoom(@RequestBody RoomNameDTO roomNameDTO) {
         springChessService.deleteRoom(roomNameDTO.getRoomName());
     }
-
-    @ExceptionHandler(NotExistRoomException.class)
-    public ResponseEntity<String> dbException(NotExistRoomException e) {
-        return ResponseEntity.badRequest().body(e.getMessage());
-    }
-
-    @ExceptionHandler(ChessException.class)
-    public ResponseEntity<String> chessException(ChessException e) {
-        return ResponseEntity.badRequest().body(e.getMessage());
-    }
-
-    @ExceptionHandler(DuplicatedRoomNameException.class)
-    public ResponseEntity<String> duplicatedRoomNameException(DuplicatedRoomNameException e) {
-        return ResponseEntity.badRequest().body(e.getMessage());
-    }
 }
